netdev-offload-tc: Introduce group ID management API
When offloading sample action to TC, userspace creates a unique ID
to map sFlow action and tunnel info and passes this ID to kernel instead
of the sFlow info. psample will send this ID and sampled packet to
userspace. Using the ID, userspace can recover the sFlow info and send
sampled packet to the right sFlow monitoring host.

Issue: 2181036
Change-Id: I282691853521bf9135a021e30ebd0971d550ccd2
Signed-off-by: Zhu Yanjun <yanjunz@nvidia.com>