Fix error handling for sample action
Remove incorrect attempt to unlock mutex when the lock is not held

Group ID is created in userspace as a unique ID to map sFlow action
and tunnel info and passes this ID to kernel instead of the sFlow
info.

Group_id will be alloced if there is a sample action and should be
freed in failure.

Issue: 2334580
Change-Id: I13d4783090592faa4e15ebd04d9f177e6d3ccecc
Fixes: eadd19347ff3 ("netdev-offload-tc: Add offload support for sFlow")
Signed-off-by: Gavin Li <gavinl@nvidia.com>