sFlow, Get netdev ignoring packet receiving interface
Sampled VxLAN packets would be dropped by OVS if they were received from an
interface index of a port which is not in use by OVS, which is the case if
the uplink was not added to an OVS bridge.

To overcome this, get the netdev from the dpif_class to get the sample attr.

Issue: 2362514
Change-Id: Ia2884de7f64913758f781c5d6fbab08ad1c7fed5
Signed-off-by: Gavin Li <gavinl@nvidia.com>
Reviewed-by: Gavi Teitz <gavi@nvidia.com>